Add Value Scala trait and update sbt.json filepath
A new Scala trait named Value has been added to the project. Additionally, the filepath for the sbt-launch jar in the sbt.json file has been updated to reflect the current IntelliJ IDEA version and location. These changes improve the readability and organization of the project's Scala classes and ensure the correct version of the sbt-launch jar is being used.